==> 01_IDAO_Preproc_datetime.ipynb <==
# 01_IDAO_Preproc_datetime

* выделение меток времени из датасетов.

## Данные

* id — integer measurement id
* epoch — datetime in "%Y-%m-%dT%H:%M:%S.%f” format (like 2014-02-01T00:44:57.685)
* sat_id — integer satellite id
* x, y, z, x_sim, y_sim, z_sim — real (obtained with the precise simulator) and simulated (obtained with SGP4-simulator) coordinates of the satellite (km)
* Vx, Vy, Vz, Vx_sim, Vy_sim, Vz_sim — real (obtained with the precise simulator) and simulated (obtained with SGP4-simulator) velocities of the satellite (km/s)

## 0. Подключение библиотек

import pandas as pd
import matplotlib.pyplot as plt

## 1. Обзор данных

# Тренировочные данные
train = pd.read_csv('data/train.csv')
test = pd.read_csv('data/Track 1/test.csv')

# Задача 1
submission = pd.read_csv('data/Track 1/submission.csv')

train.head(3)

test.head(3)

Преобразуем столбец **epoch** во временные метки:

def time_preproc(row):
    row['epoch_num'] = row['epoch'].split('.')[1]
    row['datetime'] = row['epoch'].split('.')[0]
    return row

def df_epoch_preproc(df):
    df = df.apply(time_preproc, axis=1)
    df['datetime'] = pd.to_datetime(df['datetime'], format='%Y-%m-%dT%H:%M:%S')
    return df

train = df_epoch_preproc(train)
test = df_epoch_preproc(test)

train.head(3)

test.head(3)

Удаление лишних столбцов:

train = train.drop(['epoch', 'epoch_num'], axis=1)
test = test.drop(['epoch', 'epoch_num'], axis=1)

Сохранение датасетов:

train.to_csv('data/train_proc_with_id.csv')
test.to_csv('data/Track 1/test_proc_with_id.csv')
