==> 04_IDAO_Preproc_Test_Interpolate.ipynb <==
# 04_IDAO_Preproc_Test_Interpolate

* поминутная интерполяция значений тестового датасета

import sys
import warnings
warnings.filterwarnings('ignore')

import math
from datetime import datetime, timedelta

import matplotlib.pylab as plt
import pandas as pd
import numpy as np

from scipy import interpolate

#Тренировочные данные
df_train = pd.read_csv('data/train_proc_with_id.csv', index_col='datetime', parse_dates=True)
#Тестовые данные
df_test = pd.read_csv('data/Track 1/test_proc_with_id.csv', index_col='datetime', parse_dates=True)

# Задача 1
submission = pd.read_csv('data/Track 1/submission.csv')

df_train = df_train.drop(['Unnamed: 0'], axis=1)
df_test = df_test.drop(['Unnamed: 0'], axis=1)

df_test.head(3)

Добавим столбец, переводящий время в абсолютное значение:

df_test = df_test.reset_index()
df_test['time'] = (df_test['datetime'] - np.datetime64('1970-01-01 00:00:00')) / np.timedelta64(1, 's')

Вычислим индексы с совпадающими метками времени для дальнейшей обработки:

equal_values = []
for i in range(df_test.shape[0] - 1):
    time_prv = list(df_test.loc[i, ['datetime']])
    time_now = list(df_test.loc[i + 1, ['datetime']])
    if time_prv == time_now:
        equal_values.append(i)

Исправим совпадающие временные метки при помощи уравнения равноускоренного движения:

def improve_time_equals(df, equals_list):
    amount = len(equals_list)
    count = 0
    
    for id0 in equals_list:
        id1 = id0 + 1
        id2 = id0 + 2
        
        x0 = float(df.loc[id0, ['x_sim']])
        x1 = float(df.loc[id1, ['x_sim']])
        x2 = float(df.loc[id2, ['x_sim']])

        S02 = x2 - x0

        V0 = float(df.loc[id0, ['Vx_sim']])
        V1 = float(df.loc[id1, ['Vx_sim']])
        V2 = float(df.loc[id2, ['Vx_sim']])

        t0 = float(df.loc[id0, ['time']])
        t2 = float(df.loc[id2, ['time']])
        
        T02 = t2 - t0
        
        a = 2 * (S02 - V0 * T02) / T02**2
        
        A = a / 2
        B = V0
        C = x0 - x1

        discr = B ** 2 - 4 * A * C
        if discr >= 0:
            count += 1
            t1_1 = int((-B - math.sqrt(discr)) / (2 * A))
            t1_2 = int((-B + math.sqrt(discr)) / (2 * A))
            if min(t1_1, t1_2) < 0:
                t1 = max(t1_1, t1_2)
            else:
                t1 = min(t1_1, t1_2)
            time_delta = timedelta(seconds=t1)
            df.loc[id1, ['datetime']] = df.loc[id1, ['datetime']] + time_delta
    return df

df_test_improved = improve_time_equals(df_test, equal_values)

Не все значения удалось исправить. Получим список неисправленных значений:

equal_values = []
for i in range(df_test_improved.shape[0] - 1):
    time_prv = list(df_test_improved.loc[i, ['datetime']])
    time_now = list(df_test_improved.loc[i + 1, ['datetime']])
    if time_prv == time_now:
        equal_values.append(i)

Исправим оставшиеся значения. В качестве метки времени выберем момент времени, равноудалённый от соседних точек.

def improve_time_equals_last(df, equals_list):
    for id0 in equals_list:
        id1 = id0 + 1
        id2 = id0 + 2
        time0 = df.loc[id0, ['time']]
        time2 = df.loc[id0 + 2, ['time']]
        delta = int((time2 - time0)/2)
        time_delta = timedelta(seconds=delta)
        df.loc[id1, ['datetime']] = df.loc[id1, ['datetime']] + time_delta
    return df

df_test_improved = improve_time_equals_last(df_test_improved, equal_values)

Получим список спутников из тестовой выборки:

set_test_satellites_id = set(df_test['sat_id'])

Функции, обеспечивающие интерполяцию значений:

# производит разметку по времени в выборке значений одного спутника
def resample_for_interpolate(df_one_sat):
    df_one_sat['date'] = df_one_sat.index
    df_one_sat['time'] = (df_one_sat['date'] - np.datetime64('1970-01-01 00:00:00')) / np.timedelta64(1, 's')
    df_one_sat = df_one_sat.resample('min').mean()
    df_one_sat['date'] = df_one_sat.index
    df_one_sat['time'] = (df_one_sat['date'] - np.datetime64('1970-01-01 00:00:00')) / np.timedelta64(1, 's')
    return df_one_sat

# интерполирует одну колонку датафрейма
def interpolate_col_df(df_one_sat, df_one_sat_resampled, column='x'):
    x = df_one_sat['time']
    y = df_one_sat[column]
    xnew = df_one_sat_resampled['time']
    f = interpolate.interp1d(x, y, fill_value="extrapolate")
    ynew = f(xnew)
    df_one_sat_resampled[column] = ynew
    return df_one_sat_resampled

# интерполирует весь датафрейм
def df_interpolate(df):
    df_result = pd.DataFrame(columns=['sat_id',
                                     'id', 'x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
                                     'time', 'date'])
    
    for sat in list(set_test_satellites_id):
        df_one_sat = df.query('sat_id == @sat')
        df_one_sat_resampled = resample_for_interpolate(df_one_sat)

        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'sat_id')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'x_sim')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'y_sim')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'z_sim')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'Vx_sim')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'Vy_sim')
        df_interpolated = interpolate_col_df(df_one_sat, df_one_sat_resampled, 'Vz_sim')
        
        df_result = pd.concat([df_result, df_interpolated])
    
    return df_result

Вернём датасет к исходному виду:

df_test_improved.index = df_test_improved['datetime']
df_test_improved = df_test_improved.drop(['datetime', 'time'], axis=1)

Применим интерполяцию к тестовому датасету:

df_final = df_interpolate(df_test_improved)

Проверим полученный датасет:

df_final.head(3)

df_final.info()

Сохраним в файл:

df_final.to_csv('data/Track 1/test_final.csv')
