==> 07_IDAO_Predict.ipynb <==
# 07_IDAO_Predict

* предсказываем значения на всех ранее изученных моделях
* сохраняем предсказания в форму **submission**

import sys
import warnings
warnings.filterwarnings('ignore')
import time
from sklearn.externals import joblib
import matplotlib.pylab as plt
import pandas as pd
import numpy as np

from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import make_scorer
from sklearn.preprocessing import MinMaxScaler

from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
import lightgbm as ltb

import seaborn as sns
sns.set(style="whitegrid")

features_x = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y']

features_y = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y']

features_z = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y']

features_Vx = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y'
            ]

features_Vy = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y'
            ]

features_Vz = ['x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim',
            'sum_x_y', 'sum_x_z', 'sum_y_z',
            'dif_x_y', 'dif_x_z', 'dif_y_z',
            'mul_x_y', 'mul_x_z',
            'sum_Vx_Vy', 'sum_Vx_Vz', 'sum_Vy_Vz',
            'dif_Vx_Vy',
            'mul_Vx_Vy',
            'div_x_y'
            ]

def smape(satellite_predicted_values, satellite_true_values): 
    # the division, addition and subtraction are pointwise 
    return np.mean(np.abs((satellite_predicted_values - satellite_true_values) 
        / (np.abs(satellite_predicted_values) + np.abs(satellite_true_values))))

smape_score = make_scorer(smape)

#Тестовые данные
df_test = pd.read_csv('data/Track 1/test_final.csv')

df_test = df_test.drop(['Unnamed: 0'], axis=1)

df_test = df_test.drop(['time', 'date'], axis=1)
neworder = ['id', 'sat_id', 'x_sim','y_sim','z_sim','Vx_sim','Vy_sim','Vz_sim']
df_test = df_test.reindex(columns=neworder)
df_test['sat_id'] = df_test['sat_id'].astype('int64')

sat_id_test = list(set(df_test['sat_id']))

df_test.head(3)

def calc_smape(df):
    smape_x = smape(df['x'], df['x_sim'])
    smape_y = smape(df['y'], df['y_sim'])
    smape_z = smape(df['z'], df['z_sim'])
    
    smape_Vx = smape(df['Vx'], df['Vx_sim'])
    smape_Vy = smape(df['Vy'], df['Vy_sim'])
    smape_Vz = smape(df['Vz'], df['Vz_sim'])    
    print('sMAPE x:', smape_x)
    print('sMAPE y:', smape_y)
    print('sMAPE z:', smape_z)
    print()
    
    print('sMAPE Vx:', smape_Vx)
    print('sMAPE Vy:', smape_Vy)
    print('sMAPE Vz:', smape_Vz)
    print()
    
    final_smape = (smape_x + smape_y + smape_z + smape_Vx + smape_Vy + smape_Vz) / 6
    print('smape Final', final_smape)

def features_engeneering(df_one_sat):
    df_one_sat['sum_x_y'] = df_one_sat['x_sim'] + df_one_sat['y_sim']
    df_one_sat['sum_x_z'] = df_one_sat['x_sim'] + df_one_sat['z_sim']
    df_one_sat['sum_y_z'] = df_one_sat['y_sim'] + df_one_sat['z_sim']

    df_one_sat['dif_x_y'] = df_one_sat['x_sim'] - df_one_sat['y_sim']
    df_one_sat['dif_x_z'] = df_one_sat['x_sim'] - df_one_sat['z_sim']
    df_one_sat['dif_y_z'] = df_one_sat['y_sim'] - df_one_sat['z_sim']

    df_one_sat['mul_x_y'] = df_one_sat['x_sim'] * df_one_sat['y_sim']
    df_one_sat['mul_x_z'] = df_one_sat['x_sim'] * df_one_sat['z_sim']
    df_one_sat['mul_y_z'] = df_one_sat['y_sim'] * df_one_sat['z_sim']

    df_one_sat['div_x_y'] = df_one_sat['x_sim'] / df_one_sat['y_sim']
    df_one_sat['div_x_z'] = df_one_sat['x_sim'] / df_one_sat['z_sim']
    df_one_sat['div_y_z'] = df_one_sat['y_sim'] / df_one_sat['z_sim']

    df_one_sat['sum_Vx_Vy'] = df_one_sat['Vx_sim'] + df_one_sat['Vy_sim']
    df_one_sat['sum_Vx_Vz'] = df_one_sat['Vx_sim'] + df_one_sat['Vz_sim']
    df_one_sat['sum_Vy_Vz'] = df_one_sat['Vy_sim'] + df_one_sat['Vz_sim']

    df_one_sat['dif_Vx_Vy'] = df_one_sat['Vx_sim'] - df_one_sat['Vy_sim']
    df_one_sat['dif_Vx_Vz'] = df_one_sat['Vx_sim'] - df_one_sat['Vz_sim']
    df_one_sat['dif_Vy_Vz'] = df_one_sat['Vy_sim'] - df_one_sat['Vz_sim']

    df_one_sat['mul_Vx_Vy'] = df_one_sat['Vx_sim'] * df_one_sat['Vy_sim']
    df_one_sat['mul_Vx_Vz'] = df_one_sat['Vx_sim'] * df_one_sat['Vz_sim']
    df_one_sat['mul_Vy_Vz'] = df_one_sat['Vy_sim'] * df_one_sat['Vz_sim']
    return df_one_sat

def get_predictions(df_test, sat_id_test, features_x, features_y):
    
    df_answ = df_test.copy()
    df_result = pd.DataFrame(columns=['id', 'x', 'y', 'z', 'Vx', 'Vy', 'Vz',
                                      'x_sim', 'y_sim', 'z_sim', 'Vx_sim', 'Vy_sim', 'Vz_sim'])
    
    for sat_id in sat_id_test:
        print('Sat:', sat_id)
        df_test_one_sat = df_test.query('sat_id == @sat_id')
        df_test_one_sat = features_engeneering(df_test_one_sat)
        
        features_test_x = df_test_one_sat[features_x]
        features_test_y = df_test_one_sat[features_y]
        features_test_z = df_test_one_sat[features_z]
        
        features_test_Vx = df_test_one_sat[features_Vx]
        features_test_Vy = df_test_one_sat[features_Vy]
        features_test_Vz = df_test_one_sat[features_Vz]
        
        model_path_x = 'models/minute_x/lgb_x_' + str(sat_id) + '.pkl'
        model_path_y = 'models/minute_y/lgb_y_' + str(sat_id) + '.pkl'
        model_path_z = 'models/minute_z/lgb_z_' + str(sat_id) + '.pkl'
        
        model_path_Vx = 'models/minute_Vx/lgb_Vx_' + str(sat_id) + '.pkl'
        model_path_Vy = 'models/minute_Vy/lgb_Vy_' + str(sat_id) + '.pkl'
        model_path_Vz = 'models/minute_Vz/lgb_Vz_' + str(sat_id) + '.pkl'        
        
        # load model
        model_LTB_x = joblib.load(model_path_x)
        model_LTB_y = joblib.load(model_path_y)
        model_LTB_z = joblib.load(model_path_z)
        
        model_LTB_Vx = joblib.load(model_path_Vx)
        model_LTB_Vy = joblib.load(model_path_Vy)
        model_LTB_Vz = joblib.load(model_path_Vz)  
        
        df_test_one_sat['x'] = model_LTB_x.predict(features_test_x)
        df_test_one_sat['y'] = model_LTB_y.predict(features_test_y)
        df_test_one_sat['z'] = model_LTB_z.predict(features_test_z)
        
        df_test_one_sat['Vx'] = model_LTB_Vx.predict(features_test_Vx)
        df_test_one_sat['Vy'] = model_LTB_Vy.predict(features_test_Vy)
        df_test_one_sat['Vz'] = model_LTB_Vz.predict(features_test_Vz)
        
        
        
        df_result = pd.concat([df_result, df_test_one_sat])  
    return df_result    

df_final = get_predictions(df_test, sat_id_test, features_x, features_y)

df_result = df_final[df_final['id'].notna()]

calc_smape(df_result)

columns = ['id', 'x', 'y', 'z', 'Vx', 'Vy', 'Vz']
df_result = df_result[columns]
df_result['id'] = df_result['id'].astype('int64')

df_result = df_result.reset_index()
df_result = df_result.drop(['index'], axis=1)

df_result

df_result.to_csv('data/Track 1/submission_full.csv', index=False)
